requirements.txt
colorama
py
pyfrc
pynetconsole
pynetworktables
pynetworktables2js
pytest
robotpy-ctre
robotpy-hal-base
robotpy-hal-sim
robotpy-wpilib-utilities
wpilib
pygame
coverage
